+++ src/main/kotlin/com/github/strindberg/emacsj/actions/search/ISearchReplaceRegexpAction.kt
@@ -1,0 +1,7 @@
+package com.github.strindberg.emacsj.actions.search
+
+import com.github.strindberg.emacsj.search.ISearchReplaceHandler
+import com.github.strindberg.emacsj.search.SearchType
+import com.intellij.openapi.editor.actionSystem.EditorAction
+
+class ISearchReplaceRegexpAction : EditorAction(ISearchReplaceHandler(SearchType.REGEXP)), ISearchAction